Serve requests for maven-metadata.xml files
This is necessary to avoid problems when the invoked project doesn't
specify versions for certain plugins. Note that because of the changes
for MNG-4453 this may occur if the project uses a build extension that
defines lifecycle bindings without setting default plugin versions. This
is the case for maven-bundle-plugin.

To handle this case, generate maven-metadata.xml files on the fly from
the pluginManagement entries.